requirements.txt
aiofiles==0.5.0
aniso8601==8.0.0
async-exit-stack==1.0.1
async-generator==1.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
codecov==2.1.10
coverage==5.3
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.1
graphene==2.1.8
graphql-core==3.1.2
graphql-relay==3.0.0
h11==0.11.0
httptools==0.1.1
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
orjson==3.4.1
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
promise==2.3
py==1.9.0
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.1.1
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.24.0
Rx==3.1.1
six==1.15.0
starlette==0.13.8
toml==0.10.1
ujson==4.0.1
urllib3==1.25.11
uvicorn==0.12.2
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
zipp==3.4.0
